Create separate Hero components for Home page and Portfolio page
